validate principals during role/group review api
