+++ servers/zms/src/test/java/com/yahoo/athenz/zms/ZMSImplTest.java
@@ -17132,7 +17132,7 @@ public class ZMSImplTest {
         }
 
         try {
-            RoleSystemMeta rsm = createRoleSystemMetaObject(true);
+            RoleSystemMeta rsm = ZMSTestUtils.createRoleSystemMetaObject(true);
             zmsTest.putRoleSystemMeta(ctx, "domain", "role1", "auditenabled", auditRef, rsm);
             fail();
         } catch (ResourceException ex) {
@@ -17141,7 +17141,7 @@ public class ZMSImplTest {
         }
 
         try {
-            RoleMeta rm = createRoleMetaObject(true);
+            RoleMeta rm = ZMSTestUtils.createRoleMetaObject(true);
             zmsTest.putRoleMeta(ctx, "domain", "role1", "auditenabled", rm);
             fail();
         } catch (ResourceException ex) {
@@ -17880,7 +17880,7 @@ public class ZMSImplTest {
                 "user.jane");
         zmsImpl.putRole(ctx, domainName, "Role4", auditRef, false, role4);
 
-        RoleMeta rm = createRoleMetaObject(true);
+        RoleMeta rm = ZMSTestUtils.createRoleMetaObject(true);
         rm.setReviewEnabled(true);
         rm.setDeleteProtection(true);
         rm.setMemberExpiryDays(45);
@@ -21970,16 +21970,6 @@ public class ZMSImplTest {
         assertNull(zmsImpl.getPrincipalDomain(ctx));
     }
 
-    private RoleSystemMeta createRoleSystemMetaObject(Boolean auditEnabled) {
-
-        RoleSystemMeta meta = new RoleSystemMeta();
-
-        if (auditEnabled != null) {
-            meta.setAuditEnabled(auditEnabled);
-        }
-        return meta;
-    }
-
     private GroupSystemMeta createGroupSystemMetaObject(Boolean auditEnabled) {
 
         GroupSystemMeta meta = new GroupSystemMeta();
@@ -22044,7 +22034,7 @@ public class ZMSImplTest {
                 "user.john", "user.jane");
         zmsImpl.putRole(ctx, "rolesystemmetadom1", "role1", auditRef, false, role1);
 
-        RoleSystemMeta rsm = createRoleSystemMetaObject(true);
+        RoleSystemMeta rsm = ZMSTestUtils.createRoleSystemMetaObject(true);
         zmsImpl.putRoleSystemMeta(ctx, "rolesystemmetadom1", "role1", "auditenabled", auditRef, rsm);
 
         Role resRole1 = zmsImpl.getRole(ctx, "rolesystemmetadom1", "role1", true, false, false);
@@ -22075,7 +22065,7 @@ public class ZMSImplTest {
                 "user.john", "user.jane");
         zmsImpl.putRole(ctx, "rolesystemmetadom1", "role1", auditRef, false, role1);
 
-        RoleSystemMeta rsm = createRoleSystemMetaObject(true);
+        RoleSystemMeta rsm = ZMSTestUtils.createRoleSystemMetaObject(true);
 
         try {
             zmsImpl.putRoleSystemMeta(ctx, "rolesystemmetadom1", "role1", "auditenabled", null, rsm);
@@ -22120,16 +22110,6 @@ public class ZMSImplTest {
         System.clearProperty(ZMSConsts.ZMS_PROP_PRODUCT_ID_SUPPORT);
     }
 
-    private RoleMeta createRoleMetaObject(Boolean selfServe) {
-
-        RoleMeta meta = new RoleMeta();
-
-        if (selfServe != null) {
-            meta.setSelfServe(selfServe);
-        }
-        return meta;
-    }
-
     @Test
     public void testPutRoleMeta() {
 
@@ -22145,7 +22125,7 @@ public class ZMSImplTest {
                 "user.john", "user.jane");
         zmsImpl.putRole(ctx, "rolemetadom1", "role1", auditRef, false, role1);
 
-        RoleMeta rm = createRoleMetaObject(true);
+        RoleMeta rm = ZMSTestUtils.createRoleMetaObject(true);
         rm.setMemberExpiryDays(45);
         rm.setCertExpiryMins(55);
         rm.setServiceExpiryDays(45);
@@ -22182,7 +22162,7 @@ public class ZMSImplTest {
         // if we pass a null for the expiry days (e.g. old client)
         // then we're not going to modify the value
 
-        RoleMeta rm2 = createRoleMetaObject(false);
+        RoleMeta rm2 = ZMSTestUtils.createRoleMetaObject(false);
         zmsImpl.putRoleMeta(ctx, "rolemetadom1", "role1", auditRef, rm2);
 
         resRole1 = zmsImpl.getRole(ctx, "rolemetadom1", "role1", true, false, false);
@@ -22203,7 +22183,7 @@ public class ZMSImplTest {
 
         // now let's reset to 0
 
-        RoleMeta rm3 = createRoleMetaObject(false);
+        RoleMeta rm3 = ZMSTestUtils.createRoleMetaObject(false);
         rm3.setMemberExpiryDays(0);
         rm3.setServiceExpiryDays(0);
         rm3.setGroupExpiryDays(0);
@@ -22235,7 +22215,7 @@ public class ZMSImplTest {
 
         // invalid negative values
 
-        RoleMeta rm4 = createRoleMetaObject(false);
+        RoleMeta rm4 = ZMSTestUtils.createRoleMetaObject(false);
         rm4.setMemberExpiryDays(-10);
         try {
             zmsImpl.putRoleMeta(ctx, "rolemetadom1", "role1", auditRef, rm4);
@@ -22332,10 +22312,10 @@ public class ZMSImplTest {
                 "user.john", "user.jane");
         zmsImpl.putRole(ctx, "rolemetadom1", "role1", auditRef, false, role1);
 
-        RoleSystemMeta rsm = createRoleSystemMetaObject(true);
+        RoleSystemMeta rsm = ZMSTestUtils.createRoleSystemMetaObject(true);
         zmsImpl.putRoleSystemMeta(ctx, "rolemetadom1", "role1", "auditenabled", auditRef, rsm);
 
-        RoleMeta rm = createRoleMetaObject(true);
+        RoleMeta rm = ZMSTestUtils.createRoleMetaObject(true);
         try {
             zmsImpl.putRoleMeta(ctx, "rolemetadom1", "role1", null, rm);
             fail();
@@ -22573,7 +22553,7 @@ public class ZMSImplTest {
         Role selfserverole = zmsTestInitializer.createRoleObject("testdomain1", "testrole2", null,"user.john", "user.jane");
         zmsImpl.putRole(ctx, "testdomain1", "testrole2", auditRef, false, selfserverole);
 
-        RoleMeta rm = createRoleMetaObject(true);
+        RoleMeta rm = ZMSTestUtils.createRoleMetaObject(true);
         zmsImpl.putRoleMeta(ctx, "testdomain1", "testrole2",  auditRef, rm);
 
         domain = zmsImpl.getAthenzDomain("testdomain1", false);
@@ -22596,7 +22576,7 @@ public class ZMSImplTest {
         Role auditedRole = zmsTestInitializer.createRoleObject("testdomain1", "testrole3", null,"user.john", "user.jane");
         zmsImpl.putRole(ctx, "testdomain1", "testrole3", auditRef, false, auditedRole);
 
-        RoleSystemMeta rsm = createRoleSystemMetaObject(true);
+        RoleSystemMeta rsm = ZMSTestUtils.createRoleSystemMetaObject(true);
         zmsImpl.putRoleSystemMeta(ctx, "testdomain1", "testrole3", "auditenabled", auditRef, rsm);
 
         domain = zmsImpl.getAthenzDomain("testdomain1", false);
@@ -22646,7 +22626,7 @@ public class ZMSImplTest {
         Role role1 = zmsTestInitializer.createRoleObject("testdomain1", "testrole1", null, "user.john", "user.jane");
         zmsImpl.putRole(ctx, "testdomain1", "testrole1", auditRef, false, role1);
 
-        RoleMeta rm = createRoleMetaObject(true);
+        RoleMeta rm = ZMSTestUtils.createRoleMetaObject(true);
         zmsImpl.putRoleMeta(ctx, "testdomain1", "testrole1", auditRef, rm);
 
         //switch to user.bob principal to test selfserve role membership
@@ -23103,7 +23083,7 @@ public class ZMSImplTest {
 
         Role auditedRole = zmsTestInitializer.createRoleObject(domainName, roleName, null, "user.john", "user.jane");
         zmsImpl.putRole(ctx, domainName, roleName, auditRef, false, auditedRole);
-        RoleSystemMeta rsm = createRoleSystemMetaObject(true);
+        RoleSystemMeta rsm = ZMSTestUtils.createRoleSystemMetaObject(true);
         zmsImpl.putRoleSystemMeta(ctx, domainName, roleName, "auditenabled", auditRef, rsm);
 
         Membership mbr = new Membership();
@@ -23191,7 +23171,7 @@ public class ZMSImplTest {
 
         Role auditedRole = zmsTestInitializer.createRoleObject(domainName, roleName, null, "user.john", "user.jane");
         zmsImpl.putRole(ctx, domainName, roleName, auditRef, false, auditedRole);
-        RoleSystemMeta rsm = createRoleSystemMetaObject(true);
+        RoleSystemMeta rsm = ZMSTestUtils.createRoleSystemMetaObject(true);
         zmsImpl.putRoleSystemMeta(ctx, domainName, roleName, "auditenabled", auditRef, rsm);
 
         Membership mbr = new Membership();
@@ -23276,7 +23256,7 @@ public class ZMSImplTest {
 
         Role auditedRole = zmsTestInitializer.createRoleObject("testdomain1", "testrole1", null, "user.john", "user.jane");
         zmsImpl.putRole(ctx, "testdomain1", "testrole1", auditRef, false, auditedRole);
-        RoleSystemMeta rsm = createRoleSystemMetaObject(true);
+        RoleSystemMeta rsm = ZMSTestUtils.createRoleSystemMetaObject(true);
         zmsImpl.putRoleSystemMeta(ctx, "testdomain1", "testrole1", "auditenabled", auditRef, rsm);
 
         Membership mbr = new Membership();
@@ -23436,7 +23416,7 @@ public class ZMSImplTest {
 
         Role auditedRole = zmsTestInitializer.createRoleObject("testdomain1", "testrole1", null,"user.john", "user.jane");
         zmsImpl.putRole(ctx, "testdomain1", "testrole1", auditRef, false, auditedRole);
-        RoleSystemMeta rsm = createRoleSystemMetaObject(true);
+        RoleSystemMeta rsm = ZMSTestUtils.createRoleSystemMetaObject(true);
         zmsImpl.putRoleSystemMeta(ctx, "testdomain1", "testrole1", "auditenabled", auditRef, rsm);
 
         Membership mbr = new Membership();
@@ -23501,7 +23481,7 @@ public class ZMSImplTest {
 
         Role auditedRole = zmsTestInitializer.createRoleObject("testdomain1", "testrole1", null, "user.john", "user.jane");
         zmsImpl.putRole(ctx, "testdomain1", "testrole1", auditRef, false, auditedRole);
-        RoleSystemMeta rsm = createRoleSystemMetaObject(true);
+        RoleSystemMeta rsm = ZMSTestUtils.createRoleSystemMetaObject(true);
         zmsImpl.putRoleSystemMeta(ctx, "testdomain1", "testrole1", "auditenabled", auditRef, rsm);
 
         Membership mbr = new Membership();
@@ -23594,7 +23574,7 @@ public class ZMSImplTest {
 
         Role auditedRole = zmsTestInitializer.createRoleObject("testdomain1", "testrole1", null, "user.john", "user.jane");
         zmsImpl.putRole(ctx, "testdomain1", "testrole1", auditRef, false, auditedRole);
-        RoleSystemMeta rsm = createRoleSystemMetaObject(true);
+        RoleSystemMeta rsm = ZMSTestUtils.createRoleSystemMetaObject(true);
         zmsImpl.putRoleSystemMeta(ctx, "testdomain1", "testrole1", "auditenabled", auditRef, rsm);
 
         Membership mbr = new Membership();
@@ -23705,7 +23685,7 @@ public class ZMSImplTest {
 
         Role auditedRole = zmsTestInitializer.createRoleObject("testdomain1", "testrole1", null,"user.john", "user.jane");
         zmsImpl.putRole(ctx, "testdomain1", "testrole1", auditRef, false, auditedRole);
-        RoleSystemMeta rsm = createRoleSystemMetaObject(true);
+        RoleSystemMeta rsm = ZMSTestUtils.createRoleSystemMetaObject(true);
         zmsImpl.putRoleSystemMeta(ctx, "testdomain1", "testrole1", "auditenabled", auditRef, rsm);
         RoleMeta rMeta = new RoleMeta().setDeleteProtection(true);
         zmsImpl.putRoleMeta(ctx, "testdomain1", "testrole1", auditRef, rMeta);
@@ -24120,7 +24100,7 @@ public class ZMSImplTest {
         Role selfserverole = zmsTestInitializer.createRoleObject("testdomain1", "testrole2", null,"user.john", "user.jane");
         zmsImpl.putRole(ctx, "testdomain1", "testrole2", auditRef, false, selfserverole);
 
-        RoleMeta rm = createRoleMetaObject(true);
+        RoleMeta rm = ZMSTestUtils.createRoleMetaObject(true);
         zmsImpl.putRoleMeta(ctx, "testdomain1", "testrole2",  auditRef, rm);
 
         Authority auditAdminPrincipalAuthority = new com.yahoo.athenz.common.server.debug.DebugPrincipalAuthority();
@@ -24611,400 +24591,6 @@ public class ZMSImplTest {
     }
 
     @Test
-    public void testPutRoleReviewExpiration() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        final String domainName = "role-review-dom";
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName,
-                "Role review Test Domain1", "testOrg", "user.user1");
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        Role role1 = zmsTestInitializer.createRoleObject(domainName, "role1", null,
-                "user.john", "user.jane");
-        zmsImpl.putRole(ctx, domainName, "role1", auditRef, false, role1);
-
-        Timestamp tenDaysExpiry = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(10, TimeUnit.DAYS));
-        Timestamp sixtyDaysExpiry = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(60, TimeUnit.DAYS));
-
-        Timestamp fortyFiveDaysLowerBoundExpiry = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(45, TimeUnit.DAYS));
-        Timestamp fortyFiveDaysUpperBoundExpiry = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(45, TimeUnit.DAYS) + TimeUnit.MILLISECONDS.convert(5, TimeUnit.MINUTES));
-
-        Timestamp fiftyDaysLowerBoundExpiry = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(50, TimeUnit.DAYS));
-        Timestamp fiftyDaysUpperBoundExpiry = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(50, TimeUnit.DAYS) + TimeUnit.MILLISECONDS.convert(5, TimeUnit.MINUTES));
-
-        Timestamp fiftyFiveDaysLowerBoundExpiry = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(55, TimeUnit.DAYS));
-        Timestamp fiftyFiveDaysUpperBoundExpiry = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(55, TimeUnit.DAYS) + TimeUnit.MILLISECONDS.convert(5, TimeUnit.MINUTES));
-
-        Membership mbr = zmsTestInitializer.generateMembership("role1", "user.doe", tenDaysExpiry);
-        zmsImpl.putMembership(ctx, domainName, "role1", "user.doe", auditRef, false, mbr);
-
-        Group group1 = zmsTestInitializer.createGroupObject(domainName, "group1", null);
-        zmsImpl.putGroup(ctx, domainName, "group1", auditRef, false, group1);
-
-        mbr = zmsTestInitializer.generateMembership("role1", domainName + ":group.group1", tenDaysExpiry);
-        zmsImpl.putMembership(ctx, domainName, "role1", domainName + ":group.group1", auditRef, false, mbr);
-
-        mbr = zmsTestInitializer.generateMembership("role1", "sys.auth.zms", tenDaysExpiry);
-        zmsImpl.putMembership(ctx, domainName, "role1", "sys.auth.zms", auditRef, false, mbr);
-
-        RoleMeta rm = createRoleMetaObject(true);
-        rm.setMemberExpiryDays(45);
-        rm.setServiceExpiryDays(50);
-        rm.setGroupExpiryDays(55);
-        zmsImpl.putRoleMeta(ctx, domainName, "role1", auditRef, rm);
-
-        Role inputRole = new Role().setName("role1");
-        List<RoleMember> inputMembers = new ArrayList<>();
-        inputRole.setRoleMembers(inputMembers);
-        inputMembers.add(new RoleMember().setMemberName("user.john").setActive(false));
-        inputMembers.add(new RoleMember().setMemberName("user.doe").setActive(true)
-                .setExpiration(sixtyDaysExpiry));
-        inputMembers.add(new RoleMember().setMemberName(domainName + ":group.group1").setActive(true)
-                .setExpiration(sixtyDaysExpiry));
-        inputMembers.add(new RoleMember().setMemberName("sys.auth.zms").setActive(true)
-                .setExpiration(sixtyDaysExpiry));
-        zmsImpl.putRoleReview(ctx, domainName, "role1", auditRef, false, inputRole);
-
-        Role resRole1 = zmsImpl.getRole(ctx, domainName, "role1", false, false, false);
-
-        int userChecked = 0;
-        for (RoleMember roleMember : resRole1.getRoleMembers()) {
-            switch (roleMember.getMemberName()) {
-                case "user.jane":
-                case "user.doe":
-                    userChecked += 1;
-                    assertTrue(roleMember.getExpiration().toDate().after(fortyFiveDaysLowerBoundExpiry.toDate()) && roleMember.getExpiration().toDate().before(fortyFiveDaysUpperBoundExpiry.toDate()));
-                    assertTrue(roleMember.getApproved());
-                    break;
-                case "sys.auth.zms":
-                    userChecked += 1;
-                    assertTrue(roleMember.getExpiration().toDate().after(fiftyDaysLowerBoundExpiry.toDate()) && roleMember.getExpiration().toDate().before(fiftyDaysUpperBoundExpiry.toDate()));
-                    assertTrue(roleMember.getApproved());
-                    break;
-                case domainName + ":group.group1":
-                    userChecked += 1;
-                    assertTrue(roleMember.getExpiration().toDate().after(fiftyFiveDaysLowerBoundExpiry.toDate()) && roleMember.getExpiration().toDate().before(fiftyFiveDaysUpperBoundExpiry.toDate()));
-                    assertTrue(roleMember.getApproved());
-                    break;
-            }
-        }
-        assertEquals(userChecked, 4);
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
-    public void testPutRoleReviewReviewReminder() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        final String domainName = "role-review-reminder";
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName,
-                "Role review Test Domain1", "testOrg", "user.user1");
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        Role role1 = zmsTestInitializer.createRoleObject(domainName, "role1", null,
-                "user.john", "user.jane");
-        zmsImpl.putRole(ctx, domainName, "role1", auditRef, false, role1);
-
-        Timestamp tenDaysReminder = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(10, TimeUnit.DAYS));
-        Timestamp sixtyDaysReminder = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(60, TimeUnit.DAYS));
-
-        Timestamp fortyFiveDaysLowerBoundReminder = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(45, TimeUnit.DAYS));
-        Timestamp fortyFiveDaysUpperBoundReminder = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(45, TimeUnit.DAYS) + TimeUnit.MILLISECONDS.convert(5, TimeUnit.MINUTES));
-
-        Timestamp fiftyDaysLowerBoundReminder = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(50, TimeUnit.DAYS));
-        Timestamp fiftyDaysUpperBoundReminder = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(50, TimeUnit.DAYS) + TimeUnit.MILLISECONDS.convert(5, TimeUnit.MINUTES));
-
-        Timestamp fiftyFiveDaysLowerBoundReminder = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(55, TimeUnit.DAYS));
-        Timestamp fiftyFiveDaysUpperBoundReminder = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(55, TimeUnit.DAYS) + TimeUnit.MILLISECONDS.convert(5, TimeUnit.MINUTES));
-
-        Membership mbr = zmsTestInitializer.generateMembership("role1", "user.doe", tenDaysReminder);
-        zmsImpl.putMembership(ctx, domainName, "role1", "user.doe", auditRef, false, mbr);
-
-        Group group1 = zmsTestInitializer.createGroupObject(domainName, "group1", null);
-        zmsImpl.putGroup(ctx, domainName, "group1", auditRef, false, group1);
-
-        mbr = zmsTestInitializer.generateMembership("role1", domainName + ":group.group1", tenDaysReminder);
-        zmsImpl.putMembership(ctx, domainName, "role1", domainName + ":group.group1", auditRef, false, mbr);
-
-        mbr = zmsTestInitializer.generateMembership("role1", "sys.auth.zms", tenDaysReminder);
-        zmsImpl.putMembership(ctx, domainName, "role1", "sys.auth.zms", auditRef, false, mbr);
-
-        RoleMeta rm = createRoleMetaObject(true);
-        rm.setMemberReviewDays(45);
-        rm.setServiceReviewDays(50);
-        rm.setGroupReviewDays(55);
-        zmsImpl.putRoleMeta(ctx, domainName, "role1", auditRef, rm);
-
-        Role inputRole = new Role().setName("role1");
-        List<RoleMember> inputMembers = new ArrayList<>();
-        inputRole.setRoleMembers(inputMembers);
-        inputMembers.add(new RoleMember().setMemberName("user.john").setActive(false));
-        inputMembers.add(new RoleMember().setMemberName("user.doe").setActive(true)
-                .setReviewReminder(sixtyDaysReminder));
-        inputMembers.add(new RoleMember().setMemberName(domainName + ":group.group1").setActive(true)
-                .setReviewReminder(sixtyDaysReminder));
-        inputMembers.add(new RoleMember().setMemberName("sys.auth.zms").setActive(true)
-                .setReviewReminder(sixtyDaysReminder));
-        zmsImpl.putRoleReview(ctx, domainName, "role1", auditRef, false, inputRole);
-
-        Role resRole1 = zmsImpl.getRole(ctx, domainName, "role1", false, false, false);
-
-        int userChecked = 0;
-        for (RoleMember roleMember : resRole1.getRoleMembers()) {
-            switch (roleMember.getMemberName()) {
-                case "user.jane":
-                case "user.doe":
-                    userChecked += 1;
-                    assertTrue(roleMember.getReviewReminder().toDate().after(fortyFiveDaysLowerBoundReminder.toDate()) && roleMember.getReviewReminder().toDate().before(fortyFiveDaysUpperBoundReminder.toDate()));
-                    assertTrue(roleMember.getApproved());
-                    break;
-                case "sys.auth.zms":
-                    userChecked += 1;
-                    assertTrue(roleMember.getReviewReminder().toDate().after(fiftyDaysLowerBoundReminder.toDate()) && roleMember.getReviewReminder().toDate().before(fiftyDaysUpperBoundReminder.toDate()));
-                    assertTrue(roleMember.getApproved());
-                    break;
-                case domainName + ":group.group1":
-                    userChecked += 1;
-                    assertTrue(roleMember.getReviewReminder().toDate().after(fiftyFiveDaysLowerBoundReminder.toDate()) && roleMember.getReviewReminder().toDate().before(fiftyFiveDaysUpperBoundReminder.toDate()));
-                    assertTrue(roleMember.getApproved());
-                    break;
-            }
-        }
-        assertEquals(userChecked, 4);
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
-    public void testPutRoleReviewNoChanges() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        final String domainName = "role-review-no-changes";
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName,
-                "Role review Test Domain1", "testOrg", "user.user1");
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        Role role1 = zmsTestInitializer.createRoleObject(domainName, "role1", null,
-                "user.john", "user.jane");
-        zmsImpl.putRole(ctx, domainName, "role1", auditRef, false, role1);
-
-        Timestamp tenDaysExpiry = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(10, TimeUnit.DAYS));
-        Timestamp twentyDaysExpiry = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(20, TimeUnit.DAYS));
-
-        Membership mbr = zmsTestInitializer.generateMembership("role1", "user.doe", tenDaysExpiry);
-        zmsImpl.putMembership(ctx, domainName, "role1", "user.doe", auditRef, false, mbr);
-
-        mbr = zmsTestInitializer.generateMembership("role1", "user.user1", null);
-        mbr.setReviewReminder(tenDaysExpiry);
-        zmsImpl.putMembership(ctx, domainName, "role1", "user.user1", auditRef, false, mbr);
-
-        mbr = zmsTestInitializer.generateMembership("role1", "user.user2", null);
-        mbr.setReviewReminder(tenDaysExpiry);
-        zmsImpl.putMembership(ctx, domainName, "role1", "user.user2", auditRef, false, mbr);
-
-        mbr = zmsTestInitializer.generateMembership("role1", "sys.auth.zms", tenDaysExpiry);
-        mbr.setReviewReminder(tenDaysExpiry);
-        zmsImpl.putMembership(ctx, domainName, "role1", "sys.auth.zms", auditRef, false, mbr);
-
-        Group group1 = zmsTestInitializer.createGroupObject(domainName, "group1", null);
-        zmsImpl.putGroup(ctx, domainName, "group1", auditRef, false, group1);
-
-        mbr = zmsTestInitializer.generateMembership("role1", domainName + ":group.group1", tenDaysExpiry);
-        zmsImpl.putMembership(ctx, domainName, "role1", domainName + ":group.group1", auditRef, false, mbr);
-
-        Role inputRole = new Role().setName("role1");
-        List<RoleMember> inputMembers = new ArrayList<>();
-        inputRole.setRoleMembers(inputMembers);
-        inputMembers.add(new RoleMember().setMemberName("user.john").setActive(false));
-        inputMembers.add(new RoleMember().setMemberName("user.doe").setActive(true)
-                .setExpiration(twentyDaysExpiry));
-        inputMembers.add(new RoleMember().setMemberName("user.jane").setActive(true)
-                .setExpiration(tenDaysExpiry));
-        inputMembers.add(new RoleMember().setMemberName("user.user1").setActive(true)
-                .setReviewReminder(twentyDaysExpiry));
-        inputMembers.add(new RoleMember().setMemberName("user.user2").setActive(true));
-        inputMembers.add(new RoleMember().setMemberName("sys.auth.zms").setActive(true)
-                .setReviewReminder(twentyDaysExpiry));
-        inputMembers.add(new RoleMember().setMemberName(domainName + ":group.group1").setActive(true)
-                .setReviewReminder(twentyDaysExpiry));
-
-        zmsImpl.putRoleReview(ctx, domainName, "role1", auditRef, false, inputRole);
-
-        Role resRole1 = zmsImpl.getRole(ctx, domainName, "role1", false, false, false);
-
-        // john should be deleted and all others should stay as before - no changes
-
-        int userChecked = 0;
-        for (RoleMember roleMember : resRole1.getRoleMembers()) {
-            switch (roleMember.getMemberName()) {
-                case "user.jane":
-                    assertTrue(roleMember.getApproved());
-                    assertNull(roleMember.getExpiration());
-                    assertNull(roleMember.getReviewReminder());
-                    userChecked += 1;
-                    break;
-                case "user.doe":
-                case domainName + ":group.group1":
-                    assertTrue(roleMember.getApproved());
-                    assertEquals(roleMember.getExpiration(), tenDaysExpiry);
-                    assertNull(roleMember.getReviewReminder());
-                    userChecked += 1;
-                    break;
-                case "user.user1":
-                case "user.user2":
-                    assertTrue(roleMember.getApproved());
-                    assertEquals(roleMember.getReviewReminder(), tenDaysExpiry);
-                    assertNull(roleMember.getExpiration());
-                    userChecked += 1;
-                    break;
-                case "sys.auth.zms":
-                    assertTrue(roleMember.getApproved());
-                    assertEquals(roleMember.getReviewReminder(), tenDaysExpiry);
-                    assertEquals(roleMember.getExpiration(), tenDaysExpiry);
-                    userChecked += 1;
-                    break;
-                case "user.john":
-                    fail();
-                    break;
-            }
-        }
-        assertEquals(userChecked, 6);
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
-    public void testPutRoleReviewError() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        final String domainName = "role-review-error";
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName,
-                "Role review Test Domain1", "testOrg", "user.user1");
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        Role role1 = zmsTestInitializer.createRoleObject(domainName, "role1", null,
-                "user.john", "user.jane");
-        zmsImpl.putRole(ctx, domainName, "role1", auditRef, false, role1);
-
-        Timestamp tenDaysExpiry = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(10, TimeUnit.DAYS));
-        Timestamp sixtyDaysExpiry = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(60, TimeUnit.DAYS));
-
-        Membership mbr = zmsTestInitializer.generateMembership("role1", "user.doe", tenDaysExpiry);
-        zmsImpl.putMembership(ctx, domainName, "role1", "user.doe", auditRef, false, mbr);
-
-        RoleMeta rm = createRoleMetaObject(true);
-        rm.setMemberExpiryDays(45);
-        zmsImpl.putRoleMeta(ctx, domainName, "role1", auditRef, rm);
-
-        Role inputRole = new Role().setName("role2");
-        List<RoleMember> inputMembers = new ArrayList<>();
-        inputRole.setRoleMembers(inputMembers);
-        inputMembers.add(new RoleMember().setMemberName("user.john").setActive(false));
-        inputMembers.add(new RoleMember().setMemberName("user.doe").setActive(true).setExpiration(sixtyDaysExpiry));
-
-        try {
-            zmsImpl.putRoleReview(ctx, domainName, "role1", auditRef, false, inputRole);
-            fail();
-        } catch (ResourceException re) {
-            assertEquals(re.getCode(), 400);
-        }
-
-        inputRole.setName("role1");
-        try {
-            zmsImpl.putRoleReview(ctx, "role-review-dom1", "role1", auditRef, false, inputRole);
-            fail();
-        } catch (ResourceException re) {
-            assertEquals(re.getCode(), 404);
-        }
-
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
-    public void testPutRoleReviewAuditEnabled() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        final String domainName = "role-review-audit-enabled";
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName,
-                "Role review Test Domain1", "testOrg", "user.user1");
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        Role role1 = zmsTestInitializer.createRoleObject(domainName, "role1", null,
-                "user.john", "user.jane");
-        zmsImpl.putRole(ctx, domainName, "role1", auditRef, false, role1);
-
-        Timestamp tenDaysExpiry = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(10, TimeUnit.DAYS));
-        Timestamp sixtyDaysExpiry = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(60, TimeUnit.DAYS));
-        Timestamp fortyFiveDaysLowerBoundExpiry = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(45, TimeUnit.DAYS));
-
-        Membership mbr = zmsTestInitializer.generateMembership("role1", "user.doe", tenDaysExpiry);
-        zmsImpl.putMembership(ctx, domainName, "role1", "user.doe", auditRef, false, mbr);
-
-        RoleMeta rm = createRoleMetaObject(true);
-        rm.setMemberExpiryDays(45);
-        zmsImpl.putRoleMeta(ctx, domainName, "role1", auditRef, rm);
-
-        DomainMeta meta = zmsTestInitializer.createDomainMetaObject("Domain Meta for Role review test", "NewOrg",
-                true, true, "12345", 1001);
-        zmsImpl.putDomainMeta(ctx, domainName, auditRef, meta);
-        zmsImpl.putDomainSystemMeta(ctx, domainName, "auditenabled", auditRef, meta);
-
-        RoleSystemMeta rsm = createRoleSystemMetaObject(true);
-        zmsImpl.putRoleSystemMeta(ctx, domainName, "role1", "auditenabled", auditRef, rsm);
-
-        Role inputRole = new Role().setName("role1");
-        List<RoleMember> inputMembers = new ArrayList<>();
-        inputRole.setRoleMembers(inputMembers);
-        inputMembers.add(new RoleMember().setMemberName("user.john").setActive(false));
-        inputMembers.add(new RoleMember().setMemberName("user.doe").setActive(true).setExpiration(sixtyDaysExpiry));
-        zmsImpl.putRoleReview(ctx, domainName, "role1", auditRef, false, inputRole);
-
-        Role resRole1 = zmsImpl.getRole(ctx, domainName, "role1", false, false, true);
-
-        Timestamp fortyFiveDaysUpperBoundExpiry = Timestamp.fromMillis(System.currentTimeMillis() +
-                TimeUnit.MILLISECONDS.convert(45, TimeUnit.DAYS) + TimeUnit.MILLISECONDS.convert(5, TimeUnit.MINUTES));
-
-        int userChecked = 0;
-        for (RoleMember roleMember : resRole1.getRoleMembers()) {
-            if (roleMember.getMemberName().equals("user.jane")) {
-                userChecked += 1;
-                assertTrue(roleMember.getExpiration().toDate().after(fortyFiveDaysLowerBoundExpiry.toDate())
-                        && roleMember.getExpiration().toDate().before(fortyFiveDaysUpperBoundExpiry.toDate()));
-                assertTrue(roleMember.getApproved());
-            }
-
-            // 2 records for user.doe - one approved before making the domain auditEnabled with
-            //expiry date = now + 10 and another pending as part of putRoleReview with expiry date = now + 45
-
-            if (roleMember.getMemberName().equals("user.doe")) {
-                userChecked += 1;
-                if (roleMember.getApproved() == Boolean.TRUE) {
-                    assertEquals(roleMember.getExpiration(), tenDaysExpiry);
-                } else {
-                    assertTrue(roleMember.getExpiration().toDate().after(fortyFiveDaysLowerBoundExpiry.toDate())
-                            && roleMember.getExpiration().toDate().before(fortyFiveDaysUpperBoundExpiry.toDate()));
-                }
-
-            }
-        }
-        assertEquals(userChecked, 3);
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
     public void testLoadServerPrivateKey() {
 
         ZMSImpl zmsImpl = zmsTestInitializer.getZms();
@@ -25471,7 +25057,7 @@ public class ZMSImplTest {
 
         Role auditedRole = zmsTestInitializer.createRoleObject(domainName, "testrole1", null, "user.john", "user.jane");
         zmsImpl.putRole(ctx, domainName, "testrole1", auditRef, false, auditedRole);
-        RoleSystemMeta rsm = createRoleSystemMetaObject(true);
+        RoleSystemMeta rsm = ZMSTestUtils.createRoleSystemMetaObject(true);
         zmsImpl.putRoleSystemMeta(ctx, domainName, "testrole1", "auditenabled", auditRef, rsm);
 
         Membership mbr = new Membership();
@@ -25573,7 +25159,7 @@ public class ZMSImplTest {
 
         Role auditedRole = zmsTestInitializer.createRoleObject(domainName, "testrole1", null, "user.john", "user.jane");
         zmsImpl.putRole(ctx, domainName, "testrole1", auditRef, false, auditedRole);
-        RoleSystemMeta rsm = createRoleSystemMetaObject(true);
+        RoleSystemMeta rsm = ZMSTestUtils.createRoleSystemMetaObject(true);
         zmsImpl.putRoleSystemMeta(ctx, domainName, "testrole1", "auditenabled", auditRef, rsm);
         RoleMeta rm = new RoleMeta().setSelfServe(true);
         zmsImpl.putRoleMeta(ctx, domainName, "testrole1", auditRef, rm);
@@ -28764,145 +28350,6 @@ public class ZMSImplTest {
     }
 
     @Test
-    public void testPutGroupReview() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        final String domainName = "group-review-dom";
-        final String groupName = "group1";
-
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName, "Role review Test Domain1", "testOrg", "user.user1");
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        Group group = zmsTestInitializer.createGroupObject(domainName, groupName, "user.john", "user.jane");
-        zmsImpl.putGroup(ctx, domainName, groupName, auditRef, false, group);
-
-        Group inputGroup = new Group().setName(groupName);
-        List<GroupMember> inputMembers = new ArrayList<>();
-        inputGroup.setGroupMembers(inputMembers);
-        inputMembers.add(new GroupMember().setMemberName("user.john").setActive(false));
-        inputMembers.add(new GroupMember().setMemberName("user.jane").setActive(true));
-        zmsImpl.putGroupReview(ctx, domainName, groupName, auditRef, false, inputGroup);
-
-        Group resGroup = zmsImpl.getGroup(ctx, domainName, groupName, false, false);
-        assertEquals(resGroup.getGroupMembers().size(), 1);
-        assertEquals(resGroup.getGroupMembers().get(0).getMemberName(), "user.jane");
-
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
-    public void testPutGroupReviewNoChanges() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        final String domainName = "group-review-no-changes";
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName,
-                "Role review Test Domain1", "testOrg", "user.user1");
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        Group group1 = zmsTestInitializer.createGroupObject(domainName, "group1", "user.john", "user.jane");
-        zmsImpl.putGroup(ctx, domainName, "group1", auditRef, false, group1);
-
-        Timestamp tenDaysExpiry = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(10, TimeUnit.DAYS));
-        Timestamp twentyDaysExpiry = Timestamp.fromMillis(System.currentTimeMillis() + TimeUnit.MILLISECONDS.convert(20, TimeUnit.DAYS));
-
-        GroupMembership mbr = zmsTestInitializer.generateGroupMembership("group1", "user.doe", tenDaysExpiry);
-        zmsImpl.putGroupMembership(ctx, domainName, "group1", "user.doe", auditRef, false, mbr);
-
-        mbr = zmsTestInitializer.generateGroupMembership("group1", "sys.auth.zms", tenDaysExpiry);
-        zmsImpl.putGroupMembership(ctx, domainName, "group1", "sys.auth.zms", auditRef, false, mbr);
-
-        Group inputGroup = new Group().setName("group1");
-        List<GroupMember> inputMembers = new ArrayList<>();
-        inputGroup.setGroupMembers(inputMembers);
-        inputMembers.add(new GroupMember().setMemberName("user.john").setActive(false));
-        inputMembers.add(new GroupMember().setMemberName("user.doe").setActive(true)
-                .setExpiration(twentyDaysExpiry));
-        inputMembers.add(new GroupMember().setMemberName("user.jane").setActive(true)
-                .setExpiration(tenDaysExpiry));
-        inputMembers.add(new GroupMember().setMemberName("sys.auth.zms").setActive(true)
-                .setExpiration(twentyDaysExpiry));
-
-        zmsImpl.putGroupReview(ctx, domainName, "group1", auditRef, false, inputGroup);
-
-        Group resGroup1 = zmsImpl.getGroup(ctx, domainName, "group1", false, false);
-
-        // john should be deleted and all others should stay as before - no changes
-
-        int userChecked = 0;
-        for (GroupMember groupMember : resGroup1.getGroupMembers()) {
-            switch (groupMember.getMemberName()) {
-                case "user.jane":
-                    assertTrue(groupMember.getApproved());
-                    assertNull(groupMember.getExpiration());
-                    userChecked += 1;
-                    break;
-                case "user.doe":
-                case "sys.auth.zms":
-                    assertTrue(groupMember.getApproved());
-                    assertEquals(groupMember.getExpiration(), tenDaysExpiry);
-                    userChecked += 1;
-                    break;
-                case "user.john":
-                    fail();
-                    break;
-            }
-        }
-        assertEquals(userChecked, 3);
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
-    public void testPutGroupReviewError() {
-
-        ZMSImpl zmsImpl = zmsTestInitializer.getZms();
-        RsrcCtxWrapper ctx = zmsTestInitializer.getMockDomRsrcCtx();
-        final String auditRef = zmsTestInitializer.getAuditRef();
-
-        final String domainName = "group-review-dom-err";
-        final String groupName = "group1";
-
-        TopLevelDomain dom1 = zmsTestInitializer.createTopLevelDomainObject(domainName, "Role review Test Domain1", "testOrg", "user.user1");
-        zmsImpl.postTopLevelDomain(ctx, auditRef, dom1);
-
-        Group group = zmsTestInitializer.createGroupObject(domainName, groupName, "user.john", "user.jane");
-        zmsImpl.putGroup(ctx, domainName, groupName, auditRef, false, group);
-
-        Group inputGroup = new Group().setName(groupName);
-        List<GroupMember> inputMembers = new ArrayList<>();
-        inputGroup.setGroupMembers(inputMembers);
-        inputMembers.add(new GroupMember().setMemberName("user.john").setActive(false));
-        inputMembers.add(new GroupMember().setMemberName("user.joe").setActive(true));
-        zmsImpl.putGroupReview(ctx, domainName, groupName, auditRef, false, inputGroup);
-
-        // This operation will be no-op as the changes were already implemented.
-        zmsImpl.putGroupReview(ctx, domainName, groupName, auditRef, false, inputGroup);
-
-        inputGroup.setName("group2");
-        try {
-            zmsImpl.putGroupReview(ctx, domainName, groupName, auditRef, false, inputGroup);
-            fail();
-        } catch (ResourceException re) {
-            assertEquals(re.getCode(), ResourceException.BAD_REQUEST);
-        }
-
-        inputGroup.setName(groupName);
-        try {
-            zmsImpl.putGroupReview(ctx, "invalid-domain", groupName, auditRef, false, inputGroup);
-            fail();
-        } catch (ResourceException re) {
-            assertEquals(re.getCode(), ResourceException.NOT_FOUND);
-        }
-
-        zmsImpl.deleteTopLevelDomain(ctx, domainName, auditRef);
-    }
-
-    @Test
     public void testGetPrincipalGroups() {
 
         ZMSImpl zmsImpl = zmsTestInitializer.getZms();
@@ -30113,7 +29560,7 @@ public class ZMSImplTest {
         // if we try to put the audit enabled flag on the role
         // it should be rejected since the group doesn't have audit flag
 
-        RoleSystemMeta rsm = createRoleSystemMetaObject(true);
+        RoleSystemMeta rsm = ZMSTestUtils.createRoleSystemMetaObject(true);
         try {
             zmsImpl.putRoleSystemMeta(ctx, domainName, roleName1, "auditenabled", auditRef, rsm);
             fail();
@@ -31905,7 +31352,7 @@ public class ZMSImplTest {
 
         // putRoleMeta events
         ctx = zmsTestInitializer.contextWithMockPrincipal("putRoleMeta");
-        RoleMeta rm = createRoleMetaObject(true);
+        RoleMeta rm = ZMSTestUtils.createRoleMetaObject(true);
         zmsImpl.putRoleMeta(ctx, domainName, roleName, "auditenabled", rm);
         assertSingleChangeMessage(ctx.getDomainChangeMessages(), ROLE, domainName, roleName, "putRoleMeta");
 
@@ -31960,7 +31407,7 @@ public class ZMSImplTest {
 
         // putRoleSystemMeta events
         ctx = zmsTestInitializer.contextWithMockPrincipal("putRoleSystemMeta");
-        RoleSystemMeta rsm = createRoleSystemMetaObject(true);
+        RoleSystemMeta rsm = ZMSTestUtils.createRoleSystemMetaObject(true);
         zmsImpl.putRoleSystemMeta(ctx, domainName, roleName, "auditenabled", auditRef, rsm);
         assertSingleChangeMessage(ctx.getDomainChangeMessages(), ROLE, domainName, roleName, "putRoleSystemMeta");
